# Optimize BeginScope performance