--- a/src/Microsoft.Extensions.Logging/LoggerFactory.cs
+++ b/src/Microsoft.Extensions.Logging/LoggerFactory.cs
@@ -19,8 +19,7 @@ namespace Microsoft.Extensions.Logging
         private volatile bool _disposed;
         private IDisposable _changeTokenRegistration;
         private LoggerFilterOptions _filterOptions;
-
-        internal LoggerExternalScopeProvider ScopeProvider { get; private set; }
+        private LoggerExternalScopeProvider _scopeProvider;
 
         public LoggerFactory() : this(Enumerable.Empty<ILoggerProvider>())
         {
@@ -50,15 +49,9 @@ namespace Microsoft.Extensions.Logging
             lock (_sync)
             {
                 _filterOptions = filterOptions;
-                foreach (var logger in _loggers)
+                foreach (var logger in _loggers.Values)
                 {
-                    var loggerInformation = logger.Value.Loggers;
-                    var categoryName = logger.Key;
-
-                    ApplyRules(loggerInformation, categoryName, 0, loggerInformation.Length);
-
-                    logger.Value.Loggers = loggerInformation;
-                    logger.Value.CaptureScopes = filterOptions.CaptureScopes;
+                    (logger.MessageLoggers, logger.ScopeLoggers) = ApplyFilters(logger.Loggers);
                 }
             }
         }
@@ -74,11 +67,13 @@ namespace Microsoft.Extensions.Logging
             {
                 if (!_loggers.TryGetValue(categoryName, out var logger))
                 {
-                    logger = new Logger(this)
+                    logger = new Logger
                     {
                         Loggers = CreateLoggers(categoryName),
-                        CaptureScopes = _filterOptions.CaptureScopes
                     };
+
+                    (logger.MessageLoggers, logger.ScopeLoggers) = ApplyFilters(logger.Loggers);
+
                     _loggers[categoryName] = logger;
                 }
 
@@ -97,18 +92,17 @@ namespace Microsoft.Extensions.Logging
             {
                 AddProviderRegistration(provider, dispose: true);
 
-                foreach (var logger in _loggers)
+                foreach (var existingLogger in _loggers)
                 {
-                    var loggerInformation = logger.Value.Loggers;
-                    var categoryName = logger.Key;
+                    var logger = existingLogger.Value;
+                    var loggerInformation = existingLogger.Value.Loggers;
 
+                    var newLoggerIndex = loggerInformation.Length;
                     Array.Resize(ref loggerInformation, loggerInformation.Length + 1);
-                    var newLoggerIndex = loggerInformation.Length - 1;
+                    loggerInformation[newLoggerIndex] = new LoggerInformation(provider, existingLogger.Key);
 
-                    SetLoggerInformation(ref loggerInformation[newLoggerIndex], provider, categoryName);
-                    ApplyRules(loggerInformation, categoryName, newLoggerIndex, 1);
-
-                    logger.Value.Loggers = loggerInformation;
+                    logger.Loggers = loggerInformation;
+                    (logger.MessageLoggers, logger.ScopeLoggers) = ApplyFilters(logger.Loggers);
                 }
             }
         }
@@ -123,50 +117,70 @@ namespace Microsoft.Extensions.Logging
 
             if (provider is ISupportExternalScope supportsExternalScope)
             {
-                if (ScopeProvider == null)
+                if (_scopeProvider == null)
                 {
-                    ScopeProvider = new LoggerExternalScopeProvider();
+                    _scopeProvider = new LoggerExternalScopeProvider();
                 }
 
-                supportsExternalScope.SetScopeProvider(ScopeProvider);
+                supportsExternalScope.SetScopeProvider(_scopeProvider);
             }
-        }
-
-        private void SetLoggerInformation(ref LoggerInformation loggerInformation, ILoggerProvider provider,  string categoryName)
-        {
-            loggerInformation.Logger = provider.CreateLogger(categoryName);
-            loggerInformation.ProviderType = provider.GetType();
-            loggerInformation.ExternalScope = provider is ISupportExternalScope;
         }
 
         private LoggerInformation[] CreateLoggers(string categoryName)
         {
             var loggers = new LoggerInformation[_providerRegistrations.Count];
-            for (int i = 0; i < _providerRegistrations.Count; i++)
+            for (var i = 0; i < _providerRegistrations.Count; i++)
             {
-                SetLoggerInformation(ref loggers[i], _providerRegistrations[i].Provider, categoryName);
+                loggers[i] = new LoggerInformation(_providerRegistrations[i].Provider, categoryName);
             }
-
-            ApplyRules(loggers, categoryName, 0, loggers.Length);
             return loggers;
         }
 
-        private void ApplyRules(LoggerInformation[] loggers, string categoryName, int start, int count)
+        private (MessageLogger[] MessageLoggers, ScopeLogger[] ScopeLoggers) ApplyFilters(LoggerInformation[] loggers)
         {
-            for (var index = start; index < start + count; index++)
-            {
-                ref var loggerInformation = ref loggers[index];
+            var messageLoggers = new List<MessageLogger>();
+            var scopeLoggers = _filterOptions.CaptureScopes? new List<ScopeLogger>() : null;
 
+            foreach (var loggerInformation in loggers)
+            {
                 RuleSelector.Select(_filterOptions,
                     loggerInformation.ProviderType,
-                    categoryName,
+                    loggerInformation.Category,
                     out var minLevel,
                     out var filter);
 
-                loggerInformation.Category = categoryName;
-                loggerInformation.MinLevel = minLevel;
-                loggerInformation.Filter = filter;
+                if (minLevel != null && minLevel > LogLevel.Critical)
+                {
+                    continue;
+                }
+
+                messageLoggers.Add(new MessageLogger
+                {
+                    Logger = loggerInformation.Logger,
+                    Category = loggerInformation.Category,
+                    ProviderType = loggerInformation.ProviderType,
+                    Filter = filter,
+                    MinLevel = minLevel
+                });
+
+                if (!loggerInformation.ExternalScope)
+                {
+                    scopeLoggers?.Add(new ScopeLogger
+                    {
+                        Logger = loggerInformation.Logger
+                    });
+                }
             }
+
+            if (_scopeProvider != null)
+            {
+                scopeLoggers?.Add(new ScopeLogger
+                {
+                    ExternalScopeProvider = _scopeProvider
+                });
+            }
+
+            return (messageLoggers.ToArray(), scopeLoggers?.ToArray());
         }
 
         /// <summary>