By unloading more logic from the hot path to logger creation time.

Fixes: https://github.com/aspnet/Logging/issues/894
Fixes: https://github.com/aspnet/Logging/issues/714

Also `LoggerFactory` is much more readable now.

Before:
```

               Method | CaptureScopes |     Mean |    Error |   StdDev |        Op/s |  Gen 0 | Allocated |
--------------------- |-------------- |---------:|---------:|---------:|------------:|-------:|----------:|
 MultipleMixedLoggers |         False | 185.4 ns | 1.233 ns | 1.093 ns | 5,393,087.5 |      - |       0 B |
 MultipleMixedLoggers |          True | 310.3 ns | 3.418 ns | 2.854 ns | 3,222,362.8 | 0.0005 |     168 B |
              
```
After:

```
 Method | CaptureScopes |     Mean |    Error |   StdDev |        Op/s |  Gen 0 | Allocated |
--------------------- |-------------- |---------:|---------:|---------:|------------:|-------:|----------:|
 MultipleMixedLoggers |         False | 173.6 ns | 2.438 ns | 2.281 ns | 5,761,550.5 |      - |       0 B |
 MultipleMixedLoggers |          True | 280.3 ns | 5.552 ns | 6.818 ns | 3,567,640.0 | 0.0005 |     168 B |
```